refactored websocket code to separate file, added express
