+++ package.json
@@ -12,10 +12,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "express": "^4.18.2",
     "uuid": "^9.0.0",
     "ws": "^8.11.0"
   },
   "devDependencies": {
+    "@types/express": "^4.17.15",
     "@types/node": "^18.11.15",
     "@types/uuid": "^9.0.0",
     "@types/ws": "^8.5.3",
